Implement generic datatypes
This allows things such as "data List a = ...". This required
distinguishing between types of kind Type and kinds that are of kind
"->", which will require applications to result in a type, like "List"
above, and implementing kind inference/checking. It was also necessary
to refine the parser for types to allow for applications at the type
level, with the proper precedence with respect to "->".