Switch to using a main definition for each file
Instead of simply placing an expression at the end of the file, which
would be executed. Now, the main expression may appear anywhere in the
program and is still allowed to refer to any of the other definitions.
It can also be given a type explicitly, if desired, which was not
possible before.